[AIPLATFORM-2144] chore: 중복 코드 수정
